refactor: Move containers to containers folder
